Milestones and Awards updates for Terra Cimmeria Novus (#7717)
* Explicitly specify --experimental-transform-types for ts-mocha server test, since it otherwise fails to build at least on Windows

* Milestones and Awards updates for Terra Cintra Novus based on the official expansion.

All awards were defined, just needed to be specified for this board.

For milestones, a few needed to be added.  For cases with small variations, defined alternatives within the same module as alternate exported symbols.  Introduced the notion of a TagBasedMilestone which many other milestones also fall under.  This helps eliminate defining new classes with no real behavioral/logic differences.

* Introduce notion of deprecated Milestones/Awards and deprecate unofficial/old version of some of the new ones being introduced here.

This excludes them from being randomly selected in new games (potentially along with the newer official versions in the same game) but maintains backwards compatibility for existing games and if boards explicitly specify these (as opposed to randomly select).

---------

Co-authored-by: Robert Konigsberg <kberg@users.noreply.github.com>